+++ src/config.ts
@@ -74,6 +74,7 @@ function section<T extends object>(path: string[], object: T): Settings<T> {
 export default section([root], {
 	extraEnv: value({}),
 	watchForChanges: value(true),
+	profile: value(null as string | null),
 	path: {
 		executable: value('devenv'),
 		nixBinPaths: value([] as string[]),